Update SQLFluff + pre-commit + other dependencies using requirements.in (#11)
This is an update to allow the newest SQLFluff to run in pre-commit and
for other linting tasks. It also implements a `requirements.in` file to
compile our `requirements.txt` for installing dependencies (this is
included in this PR because I did this whilst updating SQLFluff, sorry,
not very Cool Girl Atomic Commits of me).

We decided to implement a `requirements.in` because it's built-in to
[`uv`](https://github.com/astral-sh/uv) and provides a better
development experience.