[WT-7] Improve Taskfile and development flow (#17)
This PR seeks to improve the Taskfile easy setup path. In doing so, it
removes SQLFluff temporarily until a path forward for using SQFluff and
the dbt Cloud CLI can be resolved. SQLFluff's dbt templater need
dbt-core to run, which eats the dbt bin in the virtual environment that
would otherwise be the Cloud CLI. It's more important that this repo
supports the Cloud CLI for the time being, so we're removing SQLFluff.

SQL linting is still performed by pre-commit, as it installs isolated
environments in which to run all its hooks, so SQL linting on commit as
well as on demand via pre-commit commands is still available.