[Feature] Auto data generation and fixing data quality issues (#24)
This PR removes the seeds from the repo, opting for a flow that installs
the 
[`jaffle-shop-generator`](https://github.com/dbt-labs/jaffle-shop-generator) 
in the project and generates data that way. This allows more data to be
generated easily for more interesting Jaffle Shop projects with a
longer lifespan, and future proofs the repo for updates to `jafgen` (if
`jafgen` adds new entities and interactions, Jaffle Shop can build on
them instead of being stuck with the old static seeds). This PR also
fixes several data quality issues.